+++ css/idea-box.css
@@ -1,0 +1,228 @@
+* {
+  box-sizing: border-box;
+}
+body {
+  width: 100vw;
+  height: 100vh;
+  font-family: 'Open Sans', sans-serif;
+  font-weight: 400;
+}
+
+h1{
+  margin: 20px 0 20px 0;
+  font-family: 'Roboto Slab', sans-serif;
+  color: #6D6E71;
+}
+
+h2 {
+  max-width: 80%;
+  color: #6D6E71;
+}
+
+p{
+  margin: 0;
+}
+
+#title-color {
+  color: #00A79D;
+}
+
+#input-side {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 100%;
+}
+
+#enter-div {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  height: 300px;
+  width: 100%;
+  background-color: #E5F3F2;
+}
+
+#body-input {
+  width: 90%;
+  max-height: 80px;
+  min-height: 80px;
+  margin-bottom: 20px;
+  letter-spacing: 1px;
+  border: 2px solid #D1D3D4;
+  font-size: 24px;
+}
+
+#search-box{
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 100%;
+}
+
+#search,
+#title-input {
+  width: 90%;
+  margin-bottom: 20px;
+  font-size: 24px;
+  letter-spacing: 1px;
+  border: 2px solid #D1D3D4;
+}
+
+#display-side {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+}
+
+/***BUTTONS***/
+
+#save-button,
+#show-more-todos,
+#show-completed-task {
+  width: 90%;
+  height: 32px;
+  margin-bottom: 20px;
+  font-size: 20px;
+  color: #FFFFFF;
+  background: #00A79D;
+  border: none;
+  padding-bottom: 4px;
+  letter-spacing: 1px;
+}
+
+#save-button:hover {
+  background: #045E55;
+}
+
+#save-button:disabled {
+  background: #6D6E71;
+}
+
+#level-buttons {
+  margin: 0px 0 10px 10px;
+}
+
+.level-btn {
+  margin: 10px 5px;
+  display: inline-block;
+  background-color: #FFF;
+  border: none;
+}
+
+.level-btn:focus {
+  outline: 1px solid #00A79D;
+}
+
+#delete-button {
+  width: 30px;
+  margin-top: 22px;
+  background: none;
+  border: none;
+  background-image: url('../images/delete.svg');
+  background-repeat: no-repeat;
+
+}
+
+#delete-button:hover {
+  background-image: url('../images/delete-hover.svg');
+}
+
+.task-btns {
+  padding: 3px;
+  background-color: #00A79D;
+  border: none;
+  border-radius: 25px;
+  color: #ffffff;
+  font-size: 90%;
+}
+
+#upvote-button {
+  width: 30px;
+  height: 50px;
+  margin-top: 22px;
+  margin-right: 10px;
+  background: none;
+  border: none;
+  background-image: url('../images/upvote.svg');
+  background-repeat: no-repeat;
+}
+
+#upvote-button:hover {
+  background-image: url('../images/upvote-hover.svg');
+}
+
+#downvote-button {
+  width: 30px;
+  height: 50px;
+  margin-right: 10px;
+  margin-top: 22px;
+  background: none;
+  border: none;
+  background-image: url('../images/downvote.svg');
+  background-repeat: no-repeat;
+
+}
+
+#downvote-button:hover {
+  background-image: url('../images/downvote-hover.svg');
+}
+
+#completed-task {
+  width: 125px;
+  padding: 5px;
+  border: 1px solid #A9A9A9;
+}
+
+#show-more-todos {
+  width: 125px;
+  margin: 25px;
+  border: 1px solid #A9A9A9;
+  padding: 1px 5px;
+
+}
+
+/***Item-Card-Section***/
+.item-card {
+  margin: 0 auto;
+  width: 90%;
+}
+
+#card-body {
+  margin: 10px 0;
+  max-width: 90%;
+  color: #939598;
+  line-height: 25px;
+}
+
+#line-1 {
+  display: flex;
+  justify-content: space-between;
+}
+
+#line-3 {
+  display: flex;
+  justify-content: flex-start;
+  align-items: center;
+}
+
+#line-4 {
+  padding-bottom: 10px;
+  border-bottom: 2px solid #D1D3D4;
+}
+
+#quality-line {
+  color: #6D6E71;
+}
+
+/***States***/
+
+.hide {
+  display: none;
+}
+
+.completed {
+  background-color: #afebed;
+}
