+++ idea-box.html
@@ -4,8 +4,8 @@
   <head>
     <meta charset="utf-8">
     <title>idea-box</title>
-    <link rel="stylesheet" href="normalize.css">
-    <link rel="stylesheet" href="idea-box.css">
+    <link rel="stylesheet" href="css/normalize.css">
+    <link rel="stylesheet" href="css/idea-box.css">
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <link href="https://fonts.googleapis.com/css?family=Roboto+Slab:700" rel="stylesheet">
     <link href="https://fonts.googleapis.com/css?family=Open+Sans" rel="stylesheet">
@@ -41,8 +41,8 @@
     </section>
 
     <script type="text/javascript" src="http://ajax.googleapis.com/ajax/libs/jquery/3.1.1/jquery.min.js"></script>
-    <script type="text/javascript" src="jquery-3.1.1.min.js"></script>
-    <script type="text/javascript" src="idea-box.js"></script>
+    <script type="text/javascript" src="scripts/jquery-3.1.1.min.js"></script>
+    <script type="text/javascript" src="scripts/idea-box.js"></script>
 
   </body>
