+++ scripts/idea-box.js
